requirements.txt
Flask>=2.0
python-dotenv>=0.21.0
pymysql>=1.0
bcrypt>=4.0.0
cryptography>=41.0.0
pytest>=7.0.0